Add `package-info` for `paymentgateway` package in server module.
